requirements.txt
-e git+ssh://git@github.com/agiverse/agiverse-py.git@7594f6834c1841f60479db3cb975ab0adf0ab3c6#egg=agiverse
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
docutils==0.21.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
litellm==1.52.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
nh3==0.2.18
openai==1.55.0
packaging==24.2
pkginfo==1.10.0
propcache==0.2.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.21.0
sniffio==1.3.1
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.67.0
twine==5.1.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
yarl==1.18.0
zipp==3.21.0
